moved def macros to a seperate file
